Fix i18n resolving
