+++ client/src/core/session/session.resolver.ts
@@ -5,7 +5,7 @@ import { map, tap } from 'rxjs/operators';
 import { SessionModel, SessionModelSchema } from './session.model';
 
 @Injectable({ providedIn: 'root' })
-export class SessionResolver implements Resolve<any> {
+export class SessionResolver implements Resolve<SessionModel> {
 
   public session: SessionModel;
 
@@ -14,14 +14,20 @@ export class SessionResolver implements Resolve<any> {
   ) { }
 
   public resolve(): Promise<SessionModel> {
+    return this.session
+      ? Promise.resolve(this.session)
+      : this.resource();
+  }
+
+  private async resource(): Promise<SessionModel> {
     const schema = { schema: SessionModelSchema };
     return this.storage.getItem<SessionModel>('session', schema).pipe(
-      map((session) => this.refresh(session)),
+      map((session) => this.validate(session)),
       tap((session) => this.session = session)
     ).toPromise();
   }
 
-  private refresh(session: SessionModel): SessionModel {
+  private validate(session: SessionModel): SessionModel {
     const created = SessionModel.new();
     const expired = session && session.token.exp < created.token.exp;
     return !session ? created : Object.assign(session, {
